util: Replace `async-std` with `smol` (#169)
An attempt to resolve issue #71

This first commit focuses on the implementation of smol. The deprecation
of async-std would follow in a subsequent commit.

To be noted that smol does not offer an equivalent of
async_std::future::timeout so I used the select! macro together with
smol::Timer (I kept that name as such to avoid ambiguation with the
Timer type) instead.
EDIT: I found a trick to use smol's race() method instead, which I find
more readable.

I chose to implement SmolTcpListener similarly to what is done in
tokio.rs
This was harder to do for SmolUdpSocket so I sticked to the approach
taken in async_std.rs