core: Parse SAM private key correctly (#203)
Previously `SamCommand` parser assumed that the private key length of
`$privkey` was 32 bytes, corresponding to a X25519 private key. This was
incorrect as the certificate in `Destination` specified the actual
length of the (ignored) private key, causing issues with private key
files that didn't use X25519, as reported in #190.

Refactor `Destination` to contain signing and private key lengths and
use those fields to correctly parse the `$privkey` file, allowing users
to generate destination key files using i2pd-tools.

Thanks to @JohnClint-ai for root-causing the issue!

Resolves #202