cli: Change the on-disk format of address book files (#222)
Store Base32 addresses in a separate file and load the entire file in
memory when the router starts. This allows simplifying the `AddressBook`
interface as it no longer has to return a future which returns
`Option<String>` and can instead return `Option<String>` directly.

Base64-encoded destinations are stored in a separate directory, each in
its own file and the resolve logic for them remains the same.

The old address book file (`~/.emissary/addressbook/addresses`) must be
deleted in order for the new address book to function correctly